// framework/src/Models/SmartObject/extentions.ts
// Generated by https://quicktype.io

import { IExtensionConfigurationSmartObject, IExtensionSmartObject } from "../../interfaces/SmartObjects/Iextentions";
import { SmartObjectBase } from "./smartObject";

export class ExtensionConfigurationSmartObject extends SmartObjectBase<ExtensionConfigurationSmartObject> implements IExtensionConfigurationSmartObject {
    Enabled:    string = ""
    Extension:  string = ""
    ViewOrForm: string = ""
    Type:       string = ""
    Joins     = new Array<ExtensionSmartObject>();
}



export class ExtensionSmartObject extends SmartObjectBase<ExtensionSmartObject> implements  IExtensionSmartObject {
    Name:                  string = ""
    Description:           string = ""
    Type:                  string = ""
    Code:                  string = ""
    Enabled:               string = ""
    Module_Url:            string = ""
    initialiseCommand:     string = ""
    JoinedSmartObjectName: string = ""
}